refactor: fnspec -> internal.fnspec
